+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - [PR #2751](https://github.com/rqlite/rqlite/pull/2751): Add _CopyDir_ functionality to file utils.
 - [PR #2752](https://github.com/rqlite/rqlite/pull/2752): Add ability to clone Snapshots, primarily for testing purposes.
 - [PR #2753](https://github.com/rqlite/rqlite/pull/2753): `FSMFinalizer` takes a Sink.
+- [PR #2758](https://github.com/rqlite/rqlite/pull/2758): Consolidate SQLite driver construction to make connection behaviors composable. Thanks @karangupta982
 
 ## v10.2.7 (July 5th 2026)
 ### Implementation changes and bug fixes