+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v10.3.0 (unreleased)
+## v10.3.0 (September 1st 2026)
 ### New features
 - [PR #2728](https://github.com/rqlite/rqlite/pull/2728): Console app supports quickly querying for first 100 rows of a table. Thanks @rodionlim
 
@@ -19,6 +19,7 @@
 - [PR #2752](https://github.com/rqlite/rqlite/pull/2752): Add ability to clone Snapshots, primarily for testing purposes.
 - [PR #2753](https://github.com/rqlite/rqlite/pull/2753): `FSMFinalizer` takes a Sink.
 - [PR #2758](https://github.com/rqlite/rqlite/pull/2758): Consolidate SQLite driver construction to make connection behaviors composable. Thanks @karangupta982
+- [PR #2756](https://github.com/rqlite/rqlite/pull/2756): Store Index and Term in Fingerprint file and check during start-up. Fixes issue [#2747](https://github.com/rqlite/rqlite/issues/2747). Thanks @rohanpadhye
 
 ## v10.2.7 (July 5th 2026)
 ### Implementation changes and bug fixes